// src/modules/operating-room/operative-fairy-nursing-notes/models/catatan-keperawatan-pasca-operasi.ts
export interface IAktual {
    Masalah_Aktual_Teks: string;
    Time_Masalah_Aktual: string;
    Masalah_Aktual_Tindakan_Teks: string;
    Masalah_Aktual_Instruksi_Teks: string;
}

export class Aktual  {
    Masalah_Aktual_Teks: string;
    Time_Masalah_Aktual: string;
    Masalah_Aktual_Tindakan_Teks: string;
    Masalah_Aktual_Instruksi_Teks: string;
    constructor(res: IAktual) {
      this.Masalah_Aktual_Teks = res.Masalah_Aktual_Teks;
      this.Time_Masalah_Aktual = res.Time_Masalah_Aktual;
      this.Masalah_Aktual_Tindakan_Teks = res.Masalah_Aktual_Tindakan_Teks;
      this.Masalah_Aktual_Instruksi_Teks = res.Masalah_Aktual_Instruksi_Teks;
    }

    static createFromModel(json: IAktual) {
      return {
        masalah_aktual_teks: json.Masalah_Aktual_Teks ?? '',
        time_masalah_aktual: json.Time_Masalah_Aktual ?? '',
        masalah_aktual_tindakan_teks: json.Masalah_Aktual_Tindakan_Teks ?? '',
        masalah_aktual_instruksi: json.Masalah_Aktual_Instruksi_Teks ?? '',
      }
    }
}